audio: Update tone duration support from uint16 to uint32
 - Previously, tone duration was limited to the maximum value of uint16.
 - PAL supports tone duration up to the maximum value of uint32.
 - Updated tone and DTMF duration to support uint32 maximum value (treated as infinite).

CRs-Fixed: 4426758

+++ apps/samples/audio/generate_single_tone/GenerateSingleTone.cpp
@@ -170,8 +170,8 @@ int GenerateSingleTone::generateSingleTone() {
     telux::common::ErrorCode ec;
 
     std::vector<uint16_t> frequency{1244};
-    uint16_t duration = 12000;
-    uint16_t gain = 5000;
+    uint32_t duration = 12000;
+    uint16_t gain     = 5000;
 
     status = audioToneStream_->playTone(frequency, duration, gain,
         [&p] (telux::common::ErrorCode result) {
