+++ simulation/libs/audio/ICommunicator.hpp
@@ -67,7 +67,7 @@ class ICommunicator {
     virtual telux::common::Status stopStream(uint32_t streamId,
         std::shared_ptr<telux::audio::IStopStreamCb> resultListener, int cmdId) = 0;
 
-    virtual telux::common::Status playDtmfTone(telux::audio::DtmfTone dtmfTone, uint16_t duration,
+    virtual telux::common::Status playDtmfTone(telux::audio::DtmfTone dtmfTone, uint32_t duration,
         uint16_t gain, uint32_t streamId,
         std::shared_ptr<telux::audio::IDTMFCb> resultListener, int cmdId) = 0;
 
@@ -102,7 +102,7 @@ class ICommunicator {
         telux::audio::AudioUserData *audioUserData) = 0;
 
     virtual telux::common::Status playTone(uint32_t streamId, std::vector<uint16_t> frequency,
-        uint16_t duration, uint16_t gain,
+        uint32_t duration, uint16_t gain,
         std::shared_ptr<telux::audio::IToneCb> resultListener, int cmdId) = 0;
 
     virtual telux::common::Status stopTone(uint32_t streamId,
